[ez] Move AIConfigContext into a `contexts` folder
No functional changes. I think this will be a bit cleaner when we make a READ-ONLY mode because I'm going to use a context for that too and better to keep these files together
