+++ python/src/aiconfig/editor/client/src/components/AIConfigEditor.tsx
@@ -46,7 +46,7 @@ import {
 import { debounce, uniqueId } from "lodash";
 import PromptMenuButton from "./prompt/PromptMenuButton";
 import GlobalParametersContainer from "./GlobalParametersContainer";
-import AIConfigContext from "./AIConfigContext";
+import AIConfigContext from "../contexts/AIConfigContext";
 import ConfigNameDescription from "./ConfigNameDescription";
 import {
   AUTOSAVE_INTERVAL_MS,
